Recognise indexes on polymorphic associations
This commit fixes a bug that made the program to report indexes on
polymorphic associations, e.g. (commentable_type, commentable_id), as
unindexed.

Rails indexes the type column first and ID second. This confused the
algorithm which expected the ID to always be the first column in the
index.

Fix #3