Overhaul the test suite
This commit overhauls the test suite. The dummy app provided by Rails by
default turned out to be difficult to manage because of a large number
of cases. We replace it by using temping - a gem for creating temporary
tables and models.

Also, the test for UndefinedTableReferences was dropped because it
didn't play nicely with temping and is considered to be a low-risk task.