Fix missing_unique_indexes in Rails 6
Rails 6 changed the way default validator options are processed -
default values are no longer explicitly set. This caused
options[:case_sensitive] to return nil which is NOT the correct default
value for the option as it's true by default.

This commit fixes the issue by ensuring the default is provided in our
code if the key is absent from options.