Fix active_record_doctor:add_index in Rails 6
This commit fixes a bug that caused add_index to create invalid
migrations in Rails 6.0. Version numbers were added to migrations
starting with Rails 5.1. The root cause of the bug was a logical error
in the version check which expected the major version to be 5 or more
AND the minor version to be 1 or more. The latter condition isn't met
on Rails 6.0 hence the problem.