Rename tasks to detectors
This commit is a first step towards turning end-to-end tasks into
standalone classes. The old notion of "task" corresponds to a "detector"
which is an object capable of detecting problems. The new notion of
"task" is a Rake task  responsible for detecting and reporting problems.

Tasks will become a primary mean of end-to-end testing which is
currently impossible or difficult due to the fact some logic resides
inside rake task definitions.