Refactor the detector return mechanism
This commit changes the return value of detectors which used to be a
collection of problems detected _except_ undefined_table_references
which would return a pair of problems plus an option for use by the
printer.

We refactored this so that the return value of #run is always a pair
comprised of the collection and a hash of options (potentially empty).

This refactoring is a prerequisite for implementing a Rake task for
running all detectors because a standardised return value is required
for a proper assesment whether all detectors succeeded or not.